+++ src/main/java/subway/domain/line/dao/LineDao.java
@@ -1,0 +1,66 @@
+package subway.domain.line.dao;
+
+import org.springframework.dao.DataAccessException;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
+import subway.domain.line.domain.entity.LineEntity;
+
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+
+@Repository
+public class LineDao {
+    private final JdbcTemplate jdbcTemplate;
+    private final SimpleJdbcInsert insertAction;
+
+    private RowMapper<LineEntity> rowMapper = (rs, rowNum) ->
+            LineEntity.of(
+                    rs.getLong("id"),
+                    rs.getString("name"),
+                    rs.getString("color")
+            );
+
+    public LineDao(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+        this.insertAction = new SimpleJdbcInsert(jdbcTemplate)
+                .withTableName("lines")
+                .usingGeneratedKeyColumns("id");
+    }
+
+    public Long insert(LineEntity lineEntity) {
+        Map<String, Object> params = new HashMap<>();
+        params.put("id", lineEntity.getId());
+        params.put("name", lineEntity.getName());
+        params.put("color", lineEntity.getColor());
+
+        return insertAction.executeAndReturnKey(params).longValue();
+    }
+
+    public List<LineEntity> findAll() {
+        String sql = "select id, name, color from LINES";
+        return jdbcTemplate.query(sql, rowMapper);
+    }
+
+    public Optional<LineEntity> findById(Long id) {
+        String sql = "select id, name, color from LINES WHERE id = ?";
+        try {
+            return Optional.of(jdbcTemplate.queryForObject(sql, rowMapper, id));
+        } catch (DataAccessException e) {
+            return Optional.empty();
+        }
+    }
+
+    public void updateById(Long id, LineEntity newLineEntity) {
+        String sql = "update LINES set name = ?, color = ? where id = ?";
+        jdbcTemplate.update(sql, new Object[]{newLineEntity.getName(), newLineEntity.getColor(), id});
+    }
+
+    public void deleteById(Long id) {
+        jdbcTemplate.update("delete from Lines where id = ?", id);
+    }
+
+}
