+++ src/main/java/subway/domain/line/presentation/dto/LineResponse.java
@@ -1,0 +1,31 @@
+package subway.domain.line.presentation.dto;
+
+import subway.domain.line.domain.entity.LineEntity;
+
+public class LineResponse {
+    private Long id;
+    private String name;
+    private String color;
+
+    public LineResponse(Long id, String name, String color) {
+        this.id = id;
+        this.name = name;
+        this.color = color;
+    }
+
+    public static LineResponse of(LineEntity lineEntity) {
+        return new LineResponse(lineEntity.getId(), lineEntity.getName(), lineEntity.getColor());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+}
