+++ src/main/java/subway/domain/section/domain/Section.java
@@ -1,0 +1,4 @@
+package subway.domain.section.domain;
+
+public class Section {
+}
