+++ src/main/java/subway/domain/station/domain/FinalStation.java
@@ -1,0 +1,29 @@
+package subway.domain.station.domain;
+
+public class FinalStation {
+
+    final String finalUpStationName;
+    final String finalDownStationName;
+
+    private FinalStation(final String finalUpStationName, final String finalDownStationName) {
+        this.finalUpStationName = finalUpStationName;
+        this.finalDownStationName = finalDownStationName;
+    }
+
+    public static FinalStation of(final String finalUpStationName, final String finalDownStationName) {
+        return new FinalStation(finalUpStationName, finalDownStationName);
+    }
+
+    public boolean isNotFinalStation(final String stationName) {
+        return !finalUpStationName.equals(stationName) && !finalDownStationName.equals(stationName);
+    }
+
+    public boolean isFinalUpStation(final String stationName) {
+        return finalUpStationName.equals(stationName);
+    }
+
+    public boolean isFinalStation(final String stationName) {
+        return finalUpStationName.equals(stationName) || finalDownStationName.equals(stationName);
+    }
+
+}
