+++ src/main/java/subway/domain/station/domain/Station.java
@@ -1,0 +1,4 @@
+package subway.domain.station.domain;
+
+public class Station {
+}
