+++ src/main/java/subway/domain/station/exception/StationsAlreadyExistException.java
@@ -1,0 +1,14 @@
+package subway.domain.station.exception;
+
+import subway.global.error.exception.ErrorCode;
+import subway.global.error.exception.SubwayException;
+
+public class StationsAlreadyExistException extends SubwayException {
+
+    public static SubwayException THROW = new StationsAlreadyExistException();
+
+    public StationsAlreadyExistException() {
+        super(new ErrorCode(400, "입력한 역들이 모두 존재합니다."));
+    }
+
+}
