+++ src/main/java/subway/domain/station/presentation/dto/request/StationSaveRequest.java
@@ -1,0 +1,18 @@
+package subway.domain.station.presentation.dto.request;
+
+public class StationSaveRequest {
+
+    private String name;
+
+    public StationSaveRequest() {
+    }
+
+    public StationSaveRequest(final String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+}
