+++ src/main/java/subway/domain/station/presentation/dto/response/StationResponse.java
@@ -1,0 +1,25 @@
+package subway.domain.station.presentation.dto.response;
+
+import subway.domain.station.domain.entity.StationEntity;
+
+public class StationResponse {
+    private Long id;
+    private String name;
+
+    public StationResponse(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public static StationResponse of(StationEntity stationEntity) {
+        return new StationResponse(stationEntity.getId(), stationEntity.getName());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
