--- src/main/java/subway/presentation/dto/StationResponse.java
@@ -1,25 +1,0 @@
-package subway.presentation.dto;
-
-import subway.domain.entity.StationEntity;
-
-public class StationResponse {
-    private Long id;
-    private String name;
-
-    public StationResponse(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public static StationResponse of(StationEntity stationEntity) {
-        return new StationResponse(stationEntity.getId(), stationEntity.getName());
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
