+++ src/test/java/subway/dao/SectionEntityDaoTest.java
@@ -9,13 +9,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
-import subway.domain.dao.LineDao;
-import subway.domain.dao.SectionDao;
-import subway.domain.dao.StationDao;
-import subway.domain.entity.LineEntity;
-import subway.domain.entity.SectionEntity;
-import subway.domain.entity.StationEntity;
-import subway.exception.SectionNotFoundException;
+import subway.domain.line.dao.LineDao;
+import subway.domain.line.dao.SectionDao;
+import subway.domain.line.dao.StationDao;
+import subway.domain.line.domain.entity.LineEntity;
+import subway.domain.section.domain.entity.SectionEntity;
+import subway.domain.section.exception.SectionNotFoundException;
+import subway.domain.station.domain.entity.StationEntity;
 import subway.fixture.StationFixture.GangnamStation;
 import subway.fixture.StationFixture.JamsilStation;
