+++ src/test/java/subway/dao/StationEntityDaoTest.java
@@ -8,13 +8,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
-import subway.domain.dao.LineDao;
-import subway.domain.dao.SectionDao;
-import subway.domain.dao.StationDao;
-import subway.domain.entity.LineEntity;
-import subway.domain.entity.SectionEntity;
-import subway.domain.entity.StationEntity;
-import subway.exception.StationNotFoundException;
+import subway.domain.line.dao.LineDao;
+import subway.domain.line.dao.SectionDao;
+import subway.domain.line.dao.StationDao;
+import subway.domain.line.domain.entity.LineEntity;
+import subway.domain.section.domain.entity.SectionEntity;
+import subway.domain.station.domain.entity.StationEntity;
+import subway.domain.station.exception.StationNotFoundException;
 import subway.fixture.StationFixture.GangnamStation;
 import subway.fixture.StationFixture.JamsilStation;
